chore: improved comments and formatting and moving functions
